# Fix spillslot allocation to actually reuse spillslots.