@cfallin 2022-06-03
Actually, went back to version you reviewed and then did a more minimal revision based more directly on your suggestion; the `std::mem::take` in my iterator refactor is potentially moving a lot of memory and causing a slowdown.